Rename nc_goes reader to goes-imager_nc
