+++ satpy/readers/__init__.py
@@ -53,6 +53,7 @@ OLD_READER_NAMES = {
     'hrit_goes': 'goes-imager_hrit',
     'hrit_msg': 'seviri_hrit',
     'native_msg': 'seviri_native',
+    'nc_goes': 'goes-imager_nc',
     'scmi_abi_l1b': 'abi_l1b_scmi',
 }
