+++ satpy/readers/goes_imager_nc.py
